# lib/scanner: Allocate structure for final partial block